requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
chromadb
pypdf
python-docx
pytest
numpy
python-dotenv
backoff
httpx
streamlit
python-json-logger